update auction dao mapping

+++ auction-server/src/server/models/items/Art.java
@@ -1,9 +1,11 @@
 package server.models.items;
 
+import java.math.BigDecimal;
+
 public class Art extends Item {
 
-    public Art(int itemId, String name, double startingPrice,String description) {
-        super(itemId, name, startingPrice,description);
+    public Art(int itemId, String name, BigDecimal startingPrice, String description) {
+        super(itemId, name, startingPrice, description);
     }
 
     @Override
