+++ auction-server/src/server/models/items/Vehicle.java
@@ -1,5 +1,7 @@
 package server.models.items;
 
+import java.math.BigDecimal;
+
 public class Vehicle extends Item {
     // Các thuộc tính riêng (đang có sẵn)
     private String brand;
@@ -7,7 +9,7 @@ public class Vehicle extends Item {
     private String engine;
 
     // Đã SỬA: Đảo lại thứ tự (double startingPrice) lên trước (String description)
-    public Vehicle(int itemId, String name, double startingPrice, String description) {
+    public Vehicle(int itemId, String name, BigDecimal startingPrice,String description) {
         super(itemId, name, startingPrice, description);
     }
